pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "sfox-trader"
version = "0.1.0"
description = "sFOX trading utilities, strategies, and tools"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
# `src/sfox_trader/strats/new_pairs/pairs.py`, batch_scanner, etc.
research = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "numba>=0.57.0",
    "quantstats>=0.0.62",
]
# WebSocket market data (`sfox_trader.lib.sfox_ws`, OBA, get_mid_ws, …)
websocket = [
    "websocket-client>=1.6.0",
]
# Hyperliquid perp grid (`strats.mm_mgt.hyper_mgt`)
hyperliquid = [
    "hyperliquid-python-sdk>=0.22.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
